+++ classes/activities/class-activity.php
@@ -76,7 +76,7 @@ class Activity {
 	 * @return void
 	 */
 	public function save() {
-		$existing = \progress_planner()->get_query()->query_activities(
+		$existing = \progress_planner()->get_activities__query()->query_activities(
 			[
 				'category' => $this->category,
 				'type'     => $this->type,
@@ -85,10 +85,10 @@ class Activity {
 			'RAW'
 		);
 		if ( ! empty( $existing ) ) {
-			\progress_planner()->get_query()->update_activity( $existing[0]->id, $this );
+			\progress_planner()->get_activities__query()->update_activity( $existing[0]->id, $this );
 			return;
 		}
-		\progress_planner()->get_query()->insert_activity( $this );
+		\progress_planner()->get_activities__query()->insert_activity( $this );
 		\do_action( 'progress_planner_activity_saved', $this );
 	}
 
@@ -98,7 +98,7 @@ class Activity {
 	 * @return void
 	 */
 	public function delete() {
-		\progress_planner()->get_query()->delete_activity( $this );
+		\progress_planner()->get_activities__query()->delete_activity( $this );
 		\do_action( 'progress_planner_activity_deleted', $this );
 	}
