+++ classes/activities/class-maintenance.php
@@ -44,7 +44,7 @@ class Maintenance extends Activity {
 		$this->date    = new \DateTime();
 		$this->user_id = \get_current_user_id();
 
-		$existing = \progress_planner()->get_query()->query_activities(
+		$existing = \progress_planner()->get_activities__query()->query_activities(
 			[
 				'category'   => $this->category,
 				'type'       => $this->type,
@@ -54,10 +54,10 @@ class Maintenance extends Activity {
 			'RAW'
 		);
 		if ( ! empty( $existing ) ) {
-			\progress_planner()->get_query()->update_activity( $existing[0]->id, $this );
+			\progress_planner()->get_activities__query()->update_activity( $existing[0]->id, $this );
 			return;
 		}
-		\progress_planner()->get_query()->insert_activity( $this );
+		\progress_planner()->get_activities__query()->insert_activity( $this );
 		\do_action( 'progress_planner_activity_saved', $this );
 	}
