+++ classes/admin/widgets/class-activity-scores.php
@@ -69,7 +69,7 @@ final class Activity_Scores extends Widget {
 	 * @return int The score.
 	 */
 	public function get_score() {
-		$activities = \progress_planner()->get_query()->query_activities(
+		$activities = \progress_planner()->get_activities__query()->query_activities(
 			[
 				// Use 31 days to take into account
 				// the activities score decay from previous activities.
@@ -116,7 +116,7 @@ final class Activity_Scores extends Widget {
 			[
 				'label'    => \esc_html__( 'published content', 'progress-planner' ),
 				'callback' => function () {
-					$events = \progress_planner()->get_query()->query_activities(
+					$events = \progress_planner()->get_activities__query()->query_activities(
 						[
 							'start_date' => new \DateTime( '-7 days' ),
 							'category'   => 'content',
@@ -129,7 +129,7 @@ final class Activity_Scores extends Widget {
 			[
 				'label'    => \esc_html__( 'updated content', 'progress-planner' ),
 				'callback' => function () {
-					$events = \progress_planner()->get_query()->query_activities(
+					$events = \progress_planner()->get_activities__query()->query_activities(
 						[
 							'start_date' => new \DateTime( '-7 days' ),
 							'category'   => 'content',
@@ -195,7 +195,7 @@ final class Activity_Scores extends Widget {
 					}
 
 					// Get the activities.
-					$activities = \progress_planner()->get_query()->query_activities(
+					$activities = \progress_planner()->get_activities__query()->query_activities(
 						[
 							'category'   => 'content',
 							'type'       => 'publish',
