+++ classes/admin/widgets/class-suggested-tasks.php
@@ -30,7 +30,7 @@ final class Suggested_Tasks extends Widget {
 	 * @return int The score.
 	 */
 	public function get_score() {
-		$activities = \progress_planner()->get_query()->query_activities(
+		$activities = \progress_planner()->get_activities__query()->query_activities(
 			[
 				'category'   => 'suggested_task',
 				'start_date' => \DateTime::createFromFormat( 'Y-m-d', \gmdate( 'Y-m-01' ) ),
