+++ classes/badges/class-badge-maintenance.php
@@ -39,7 +39,7 @@ abstract class Badge_Maintenance extends Badge {
 				'priority'    => 'low',
 				'evaluate'    => function ( $goal_object ) {
 					return count(
-						\progress_planner()->get_query()->query_activities(
+						\progress_planner()->get_activities__query()->query_activities(
 							[
 								'start_date' => $goal_object->get_details()['start_date'],
 								'end_date'   => $goal_object->get_details()['end_date'],
