+++ classes/class-rest-api-stats.php
@@ -81,7 +81,7 @@ class Rest_API_Stats {
 
 		// Get the number of activities in the past week.
 		$data['activities'] = count(
-			\progress_planner()->get_query()->query_activities(
+			\progress_planner()->get_activities__query()->query_activities(
 				[
 					'start_date' => new \DateTime( '-7 days' ),
 				]
@@ -118,7 +118,7 @@ class Rest_API_Stats {
 		$scores = \progress_planner()->get_ui__chart()->get_chart_data(
 			[
 				'items_callback' => function ( $start_date, $end_date ) {
-					return \progress_planner()->get_query()->query_activities(
+					return \progress_planner()->get_activities__query()->query_activities(
 						[
 							'start_date' => $start_date,
 							'end_date'   => $end_date,
