+++ tests/phpunit/test-class-activity.php
@@ -45,7 +45,7 @@ class Activity_Test extends \WP_UnitTestCase {
 	public function test_save() {
 		$this->activity->save();
 
-		$activity = \progress_planner()->get_query()->query_activities(
+		$activity = \progress_planner()->get_activities__query()->query_activities(
 			[
 				'category' => $this->activity->category,
 				'type'     => $this->activity->type,
