+++ tests/phpunit/test-class-upgrade-migrations-111.php
@@ -20,8 +20,8 @@ class Upgrade_Migrations_111_Test extends \WP_UnitTestCase {
 	public function test_dataset_1() {
 
 		// Delete all activities.
-		\progress_planner()->get_query()->delete_activities(
-			\progress_planner()->get_query()->query_activities(
+		\progress_planner()->get_activities__query()->delete_activities(
+			\progress_planner()->get_activities__query()->query_activities(
 				[
 					'category' => 'suggested_task',
 				]
@@ -220,7 +220,7 @@ class Upgrade_Migrations_111_Test extends \WP_UnitTestCase {
 		// Create a new activity for each item.
 		foreach ( $migration_map as $old_task_id => $item ) {
 			// Check if the activity already exists.
-			$activity = \progress_planner()->get_query()->query_activities(
+			$activity = \progress_planner()->get_activities__query()->query_activities(
 				[
 					'data_id' => $old_task_id,
 				]
@@ -269,7 +269,7 @@ class Upgrade_Migrations_111_Test extends \WP_UnitTestCase {
 
 		// Verify that every value in the $items array has it's own activity.
 		foreach ( $migration_map as $item ) {
-			$activity = \progress_planner()->get_query()->query_activities(
+			$activity = \progress_planner()->get_activities__query()->query_activities(
 				[
 					'data_id' => $item['task_id'],
 				]
